fix: Move auth pages under main layout and remove the header and nav for them
