+++ app/(main)/@header/create-account/page.tsx
@@ -1,0 +1,3 @@
+export default function CreateAccountPageHeader() {
+  return null;
+}
