+++ app/(main)/@header/login/page.tsx
@@ -1,0 +1,3 @@
+export default function LoginPageHeader() {
+  return null;
+}
