+++ app/(main)/@header/send-verification/page.tsx
@@ -1,0 +1,3 @@
+export default function SendVerificationPageHeader() {
+  return null;
+}
