+++ app/(main)/@nav/create-account/page.tsx
@@ -1,0 +1,3 @@
+export default function CreateAccountPageNav() {
+  return null;
+}
