+++ app/(main)/@nav/send-verification/page.tsx
@@ -1,0 +1,3 @@
+export default function SendVerificationPageNav() {
+  return null;
+}
