+++ app/(main)/layout.tsx
@@ -1,7 +1,6 @@
 import "@/app/globals.css";
 import type { Metadata } from "next";
 import localFont from "next/font/local";
-import MainNav from "@/components/layout-ui/navigation/main-nav";
 import { Toaster } from "sonner";
 import { AppProvider } from "@/components/providers";
 import { getSession } from "@/features/auth";
@@ -25,9 +24,10 @@ export const metadata: Metadata = {
 interface RootLayoutProps {
   children: React.ReactNode;
   header: React.ReactNode;
+  nav: React.ReactNode;
 }
 
-export default async function RootLayout({ children, header }: RootLayoutProps) {
+export default async function RootLayout({ children, header, nav }: RootLayoutProps) {
   const session = await getSession();
   
   return (
@@ -38,9 +38,7 @@ export default async function RootLayout({ children, header }: RootLayoutProps) 
       <body className={`${geistSans.variable} ${geistMono.variable}`}>
         <AppProvider session={session}>
           <div className="flex min-h-screen w-full flex-col bg-muted/40">
-            <aside className="fixed inset-y-0 left-0 z-10 hidden w-14 flex-col border-r bg-background sm:flex">
-              <MainNav />
-            </aside>
+            {nav}
             <div className="flex flex-col sm:gap-4 sm:py-4 sm:pl-14">
               {header}
               <main className="grid flex-1 items-start gap-4 p-4 sm:px-6 sm:py-0 md:gap-8">
